apps/worker/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
rq==1.15.1
redis==5.0.1
psycopg2-binary==2.9.9
pydantic==2.5.2
requests==2.31.0
python-dotenv==1.0.0
gitpython==3.1.40
semgrep==1.50.0
# trufflehog3==3.0.10 - removed due to dependency conflict with semgrep
python-socketio==5.10.0
aiohttp==3.9.1 